+++ src/core/plugins/character-creator/shared/events.ts
@@ -5,4 +5,7 @@ export const MAIN_CHARACTER_CREATOR_EVENTS = {
     UPDATE: `${PAGE_NAME}:UPDATE`,
     DONE: `${PAGE_NAME}:Done`,
     EXIT: `${PAGE_NAME}:Exit`,
+    GET_OVERLAY_COUNT: `${PAGE_NAME}:GetOverCount`,
+    GET_PRIMARY_COLOR_LIST: `${PAGE_NAME}:GetPrimaryColorList`,
+    GET_SECONDARY_COLOR_LIST: `${PAGE_NAME}:GetSecondaryColorList`,
 };