+++ src/core/plugins/character-creator/webview/pages/CCBody.vue
@@ -6,9 +6,9 @@
             </div>
         </div>
         <div class="flex flex-col bg-neutral-900 rounded-md box-border text-neutral-200 font-semibold z-20 mt-2">
-            <div class="flex flex-col select-none flex-grow rounded m-2 text-center gap-2">
+            <div class="flex flex-col select-none flex-grow rounded m-2 text-center gap-3">
                 <!-- Body Type / Sex -->
-                <div class="flex flex-row justify-between items-center p-1 border-b pb-3 border-neutral-700">
+                <div class="flex flex-row justify-between items-center p-1 border-b pb-3 pt-2 border-neutral-700">
                     <div class="w-24 pr-4 text-left text-xs">Body</div>
                     <div class="flex flex-row justify-between items-center flex-grow">
                         <MenuButton @click="updateSex">