+++ src/core/plugins/character-select/webview/CharSelect.vue
@@ -2,29 +2,30 @@
     <div
         class="flex flex-col fixed right-0 bottom-0 mr-4 mb-4 bg-neutral-900 w-1/6 rounded-md box-border text-neutral-200 font-semibold z-20 text-xs opacity-90"
     >
-        <div v-if="isDeleting" class="flex flex-col select-none flex-grow rounded m-2 text-center gap-2">
-            <div
-                @click="deleteCharacterForSure"
-                class="transition-all active:scale-95 flex flex-row justify-between items-cente border text-red-400 border-red-400 hover:text-red-400 cursor-pointer bg-neutral-800 hover:bg-neutral-700 rounded-sm p-4 pt-2 pb-2"
-            >
-                <span>Delete '{{ characterName }}'</span>
+        <div v-auto-animate>
+            <div v-if="isDeleting" class="flex flex-col select-none flex-grow rounded m-2 text-center gap-2">
+                <div
+                    @click="deleteCharacterForSure"
+                    class="transition-all active:scale-95 flex flex-row justify-between items-cente border text-red-400 border-red-400 hover:text-red-400 cursor-pointer bg-neutral-800 hover:bg-neutral-700 rounded-sm p-4 pt-2 pb-2"
+                >
+                    <span>Delete '{{ characterName }}'</span>
+                </div>
+                <div
+                    @click="doNotDelete"
+                    class="transition-all hover:text-neutral-200 active:scale-95 flex flex-row justify-between items-center border border-neutral-700 cursor-pointer bg-neutral-800 hover:bg-neutral-700 rounded-sm p-4 pt-2 pb-2"
+                >
+                    <Icon :size="18" icon="icon-chevron-left" />
+                    <span>Go Back</span>
+                </div>
             </div>
-            <div
-                @click="doNotDelete"
-                class="transition-all hover:text-neutral-200 active:scale-95 flex flex-row justify-between items-center border border-neutral-700 cursor-pointer bg-neutral-800 hover:bg-neutral-700 rounded-sm p-4 pt-2 pb-2"
-            >
-                <Icon :size="18" icon="icon-chevron-left" />
-                <span>Go Back</span>
-            </div>
-        </div>
-
-        <div v-else class="flex flex-col select-none flex-grow rounded m-2 text-center gap-2">
-            <div
-                @click="deleteCharacter"
-                class="transition-all hover:text-neutral-200 active:scale-95 flex flex-row justify-between items-center border border-neutral-700 cursor-pointer bg-neutral-800 hover:bg-neutral-700 rounded-sm p-4 pt-2 pb-2"
-            >
-                <span>Delete</span>
-                <Icon :size="18" icon="icon-remove" />
+            <div v-else class="flex flex-col select-none flex-grow rounded m-2 text-center gap-2">
+                <div
+                    @click="deleteCharacter"
+                    class="transition-all hover:text-neutral-200 active:scale-95 flex flex-row justify-between items-center border border-neutral-700 cursor-pointer bg-neutral-800 hover:bg-neutral-700 rounded-sm p-4 pt-2 pb-2"
+                >
+                    <span>Delete</span>
+                    <Icon :size="18" icon="icon-remove" />
+                </div>
             </div>
         </div>
     </div>